Move object resolution code into separate class.
TODO:
* Currently just exposes the resulting maps - change to methods for
object by object or keyset / value set access.
* Tests for ObjectResolver
